chore:640 /team/service/get.tsを/team/service/fetchTeam.tsに統合した
