+++ packages/kcms/src/team/service/fetchTeam.ts
@@ -1,6 +1,6 @@
 import { Option, Result } from '@mikuroxina/mini-fn';
-import { TeamRepository } from '../models/repository';
-import { Team, TeamID } from '../models/team';
+import { TeamRepository } from '../models/repository.js';
+import { Team, TeamID } from '../models/team.js';
 
 export class FetchTeamService {
   private readonly repository: TeamRepository;
@@ -9,11 +9,25 @@ export class FetchTeamService {
     this.repository = repository;
   }
 
-  async handle(teamID: TeamID): Promise<Result.Result<Error, Team>> {
-    const teamRes = await this.repository.findByID(teamID);
-    if (Option.isNone(teamRes)) {
-      return Result.err(new Error('Team not found'));
+  async findAll(): Promise<Result.Result<Error, Team[]>> {
+    return await this.repository.findAll();
+  }
+
+  async findByID(id: TeamID): Promise<Result.Result<Error, Team>> {
+    const res = await this.repository.findByID(id);
+    if (Option.isNone(res)) {
+      return Result.err(new Error('Not found'));
     }
-    return Result.ok(Option.unwrap(teamRes));
+
+    return Result.ok(Option.unwrap(res));
+  }
+
+  async findByTeamName(name: string): Promise<Result.Result<Error, Team>> {
+    const res = await this.repository.findByTeamName(name);
+    if (Option.isNone(res)) {
+      return Result.err(new Error('Not found'));
+    }
+
+    return Result.ok(Option.unwrap(res));
   }
 }
