+++ packages/kcmsf/src/pages/result.tsx
@@ -195,4 +195,3 @@ const PreResultColum = (props: { match: PreMatch }) => {
     </>
   );
 };
-
